+++ packages/ember-validations/lib/main.js
@@ -4,4 +4,3 @@ require('ember-validations/errors');
 require('ember-validations/mixin');
 require('ember-validations/patterns');
 require('ember-validations/validators');
-require('ember-validations/ext');