+++ packages/ember-validations/lib/validators/length.js
@@ -18,7 +18,13 @@ Ember.Validations.validators.local.Length = Ember.Validations.validators.Base.ex
       }
     }
 
-    this.tokenizedLength = new Function('value', 'return (value || "").' + (this.options.tokenizer || 'split("")') + '.length');
+    this.options.tokenizer = this.options.tokenizer || function(value) { return value.split(''); };
+    // if (typeof(this.options.tokenizer) === 'function') {
+      // debugger;
+      // // this.tokenizedLength = new Function('value', 'return '
+    // } else {
+      // this.tokenizedLength = new Function('value', 'return (value || "").' + (this.options.tokenizer || 'split("")') + '.length');
+    // }
   },
   CHECKS: {
     'is'      : '==',
@@ -38,10 +44,10 @@ Ember.Validations.validators.local.Length = Ember.Validations.validators.Base.ex
     }
   },
   messageKeys: function() {
-    return Object.keys(this.MESSAGES);
+    return Ember.keys(this.MESSAGES);
   },
   checkKeys: function() {
-    return Object.keys(this.CHECKS);
+    return Ember.keys(this.CHECKS);
   },
   renderMessageFor: function(key) {
     var options = {count: this.getValue(key)}, _key;
@@ -72,7 +78,7 @@ Ember.Validations.validators.local.Length = Ember.Validations.validators.Base.ex
           continue;
         }
 
-        fn = new Function('return ' + this.tokenizedLength(this.model.get(this.property)) + ' ' + operator + ' ' + this.getValue(key));
+        fn = new Function('return ' + this.options.tokenizer(this.model.get(this.property)).length + ' ' + operator + ' ' + this.getValue(key));
         if (!fn()) {
           this.errors.pushObject(this.renderMessageFor(key));
         }