+++ packages/ember-validations/tests/errors_test.js
@@ -13,6 +13,9 @@ module('Errors test', {
         }
       }
     });
+  },
+  teardown: function() {
+    delete Ember.I18n;
   }
 });
 
@@ -56,4 +59,42 @@ test('when errors are resolved', function() {
   ok(user.get('isValid'));
   deepEqual(user.get('errors.name'), []);
   deepEqual(user.get('errors.age'), []);
+});
+
+test('validations use Ember.I18n.t to render the message if Ember.I18n is present', function() {
+  Ember.I18n = {
+    translations: {
+      errors: {
+        blank: 'muss ausgefüllt werden',
+        notANumber: 'ist keine Zahl'
+      }
+    },
+    lookupKey: function(key, hash) {
+      var firstKey, idx, remainingKeys;
+
+      if (hash[key] !== null && hash[key] !== undefined) { return hash[key]; }
+
+      if ((idx = key.indexOf('.')) !== -1) {
+        firstKey = key.substr(0, idx);
+        remainingKeys = key.substr(idx + 1);
+        hash = hash[firstKey];
+        if (hash) { return Ember.I18n.lookupKey(remainingKeys, hash); }
+      }
+    },
+    t: function(key, context) {
+      return Ember.I18n.lookupKey(key, Ember.I18n.translations);
+    }
+  };
+  
+  Ember.run(function() {
+    user = User.create();
+  });
+  equal(user.get('isValid'), false);
+  deepEqual(user.get('errors.name'), ['muss ausgefüllt werden']);
+  deepEqual(user.get('errors.age'), ['muss ausgefüllt werden', 'ist keine Zahl']);
+  Ember.run(function() {
+    user.set('age', 'thirty three');
+  });
+  equal(user.get('isValid'), false);
+  deepEqual(user.get('errors.age'), ['ist keine Zahl']);
 });