+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 ## 1.0.0.beta.2
+
+* Addition of URL validator.
+* Support i18n
+* Fixes tokenized length function
+* Fixes missing default numericality message